SQL file updated
Updated the Conversions table to use DECIMAL type for Fahrenheit and Celsius. Added a DROP TABLE statement at the end.

+++ SQL-Queries.sql
@@ -2,10 +2,11 @@ SELECT * FROM dbo.Conversions;
 
 CREATE TABLE Conversions (
      ID         INT IDENTITY(1, 1) PRIMARY KEY,
-     Fahrenheit FLOAT NOT NULL,
-     Celsius    FLOAT NOT NULL,
+     Fahrenheit        DECIMAL(18, 2) NOT NULL,
+     Celsius           DECIMAL(18, 2) NOT NULL
 );
 
 INSERT INTO dbo.Conversions (Fahrenheit, Celsius) VALUES (50, 10);
+
 
 DROP TABLE dbo.Conversions;
